requirements.txt
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
google-api-core==2.0.1
google-api-python-client==2.23.0
google-auth==2.2.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
greenlet==1.1.2
httplib2==0.19.1
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
requests==2.26.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.25
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.7
Werkzeug==2.0.1
zipp==3.6.0
